添加封包管理器和消息发送器

+++ src/main/java/top/spco/events/CommandEvents.java
@@ -49,7 +49,7 @@ public class CommandEvents {
          * @param message 原始消息
          * @param time    命令发送的时间
          */
-        void onCommand(Bot bot, Interactive from, User sender, MessageChain message, int time);
+        void onCommand(Bot bot, Interactive from, User sender, MessageChain message, long time);
     }
 
     /**
@@ -71,7 +71,7 @@ public class CommandEvents {
          * @param message    原始消息
          * @param time       命令发送的时间
          */
-        void onPrivateCommand(Bot bot, User interactor, MessageChain message, int time);
+        void onPrivateCommand(Bot bot, User interactor, MessageChain message, long time);
     }
 
     /**
@@ -94,6 +94,6 @@ public class CommandEvents {
          * @param time    命令发送的时间
          * @param message 原始消息
          */
-        void onGroupCommand(Bot bot, Group from, User sender, MessageChain message, int time);
+        void onGroupCommand(Bot bot, Group from, User sender, MessageChain message, long time);
     }
 }
