+++ src/main/java/top/spco/service/chat/Chat.java
@@ -148,7 +148,7 @@ public class Chat {
      * @param message 收到的消息
      * @param time    时间戳
      */
-    public void handleMessage(Bot bot, Interactive source, Interactive sender, MessageChain message, int time) {
+    public void handleMessage(Bot bot, Interactive source, Interactive sender, MessageChain message, long time) {
         if (stopped) {
             return;
         }
