+++ src/main/java/top/spco/service/command/CommandDispatcher.java
@@ -145,7 +145,7 @@ public class CommandDispatcher extends SimpleFeatureManager<Command> {
         });
     }
 
-    private void callCommand(Map<String, Command> targetCommands, Interactive from, User sender, MessageChain message, Bot bot, int time) {
+    private void callCommand(Map<String, Command> targetCommands, Interactive from, User sender, MessageChain message, Bot bot, long time) {
         Parser input = new Parser(message, message.toMessageContext().substring(1));
         while (input.canRead() && input.peek() != ' ') {
             input.skip();
@@ -163,7 +163,7 @@ public class CommandDispatcher extends SimpleFeatureManager<Command> {
         callCommand(targetCommands, from, sender, message, bot, time, label, input);
     }
 
-    private void callCommand(Map<String, Command> targetCommands, Interactive from, User sender, MessageChain message, Bot bot, int time, String label, Parser parser) {
+    private void callCommand(Map<String, Command> targetCommands, Interactive from, User sender, MessageChain message, Bot bot, long time, String label, Parser parser) {
         try {
             // 获取命令实例和发送者和发送者的用户实例
             Command command = targetCommands.get(label);
