+++ src/main/java/top/spco/service/command/commands/AdminCommand.java
@@ -45,7 +45,7 @@ public class AdminCommand extends GroupAbstractCommand {
     }
 
     @Override
-    public void onCommand(Bot bot, Interactive from, User sender, BotUser user, MessageChain message, int time, CommandMeta meta, String usageName) throws CommandSyntaxException {
+    public void onCommand(Bot bot, Interactive from, User sender, BotUser user, MessageChain message, long time, CommandMeta meta, String usageName) throws CommandSyntaxException {
         // TODO: 修复这个
 //        Group group = (Group) from;
 //        InteractiveList<NormalMember<?>> members = group.getMembers();
