+++ src/main/java/top/spco/service/command/commands/ModuleCommand.java
@@ -68,7 +68,7 @@ public final class ModuleCommand extends AbstractCommand {
     }
 
     @Override
-    public void onCommand(Bot bot, Interactive from, User sender, BotUser user, MessageChain message, int time, CommandMeta meta, String usageName) {
+    public void onCommand(Bot bot, Interactive from, User sender, BotUser user, MessageChain message, long time, CommandMeta meta, String usageName) {
         switch (usageName) {
             case "获取所有已加载的模块" -> {
                 var modules = SpCoBot.getInstance().moduleManager.getAll();
