+++ src/main/java/top/spco/service/command/commands/QuoteCommand.java
@@ -50,7 +50,7 @@ public class QuoteCommand extends AbstractCommand {
     }
 
     @Override
-    public void onCommand(Bot bot, Interactive from, User sender, BotUser user, MessageChain message, int time, CommandMeta meta, String usageName) throws CommandSyntaxException {
+    public void onCommand(Bot bot, Interactive from, User sender, BotUser user, MessageChain message, long time, CommandMeta meta, String usageName) throws CommandSyntaxException {
         var quote = SpCoBot.getInstance().getMessageService().getQuote(message);
         if (quote == null) {
             from.quoteReply(message, "该消息没有引用一条消息");
