+++ src/main/java/top/spco/service/command/commands/RecallCommand.java
@@ -52,7 +52,7 @@ public class RecallCommand extends GroupAbstractCommand {
     }
 
     @Override
-    public void onCommand(Bot bot, Interactive from, User sender, BotUser user, MessageChain message, int time, CommandMeta meta, String usageName) throws CommandSyntaxException {
+    public void onCommand(Bot bot, Interactive from, User sender, BotUser user, MessageChain message, long time, CommandMeta meta, String usageName) throws CommandSyntaxException {
         if (PermissionsValidator.isMemberAdmin(from, user, message)) {
             ImmutablePair<MessageSource, Message> quote = SpCoBot.getInstance().getMessageService().getQuote(message);
             if (quote == null) {
