+++ src/main/java/top/spco/service/command/commands/ReportCommand.java
@@ -58,7 +58,7 @@ public class ReportCommand extends AbstractCommand {
     }
 
     @Override
-    public void onCommand(Bot bot, Interactive from, User sender, BotUser user, MessageChain message, int time, CommandMeta meta, String usageName) throws CommandSyntaxException {
+    public void onCommand(Bot bot, Interactive from, User sender, BotUser user, MessageChain message, long time, CommandMeta meta, String usageName) throws CommandSyntaxException {
         from.quoteReply(message, SpCoBot.getInstance().getRuntimeStatistic().genReport());
     }
 }
