+++ src/main/java/top/spco/service/command/commands/SignCommand.java
@@ -44,7 +44,7 @@ public class SignCommand extends AbstractCommand {
     }
 
     @Override
-    public void onCommand(Bot bot, Interactive from, User sender, BotUser user, MessageChain message, int time, CommandMeta meta, String usageName) {
+    public void onCommand(Bot bot, Interactive from, User sender, BotUser user, MessageChain message, long time, CommandMeta meta, String usageName) {
         sign(from, user, message);
     }
